pyproject.toml
[project]
name = "madr-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "lucas.carvalho",email = "lucascarvalho45054@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "types-factory-boy (>=0.4.1,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev madr_fastapi/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr_fastapi -vv'
post_test = 'coverage html'